+++ src/App/KeyedEachBlocks.fs
@@ -31,7 +31,7 @@ let ThingView (viewId : int) (thing : IObservable<Thing>) : SutilElement =
         ]
 
         Html.div [
-            Bind.fragment thing <| fun t ->
+            Bind.el thing <| fun t ->
                 Html.p [
                     Html.span [ style [ Css.backgroundColor t.Color ]; text $"{t.Id} {t.Color} #{viewId}" ]
                     Html.span [ style [ Css.backgroundColor initialColor ]; text "initial" ]