+++ src/App/Markdown.fs
@@ -13,6 +13,19 @@ let style = [
     rule "ul" [ Css.allRevert ]
     rule "li" [ Css.allRevert ]
     rule "p" [ Css.marginTop (em 1); Css.marginBottom (em 1)]
+    rule "code.more" [
+        Css.maxHeight (rem 12)
+        Css.overflowYHidden
+        Css.custom("mask-image", "linear-gradient(to bottom, black 50%, transparent 100%)")
+        Css.transition "max-height 0.5s cubic-bezier(0, 1, 0, 1)"
+     ]
+    rule "code.full" [
+        Css.maxHeight (px 1000)
+        Css.transition "max-height 1s ease-in-out"
+    ]
+    rule "code.full .more-button" [
+        Css.displayNone
+    ]
 ]
 
 let withMarkdown s =