+++ src/App/Todos.fs
@@ -227,7 +227,7 @@ let view () : SutilElement =
 
         Html.div [
             class' "complete-all-container"
-            Bind.fragment model <| fun m -> Html.a [
+            Bind.el model <| fun m -> Html.a [
                 Attr.href "#"
                 text "toggle sort"
                 onClick (fun _ -> not m.Sort |> SetSort |> dispatch) [ PreventDefault ]
@@ -239,7 +239,7 @@ let view () : SutilElement =
             ]
             Html.span [
                 class' "kudos"
-                Bind.fragment completed (fun x -> text $"{x.Length} tasks completed! Good job!")
+                Bind.el completed (fun x -> text $"{x.Length} tasks completed! Good job!")
             ] |> fader lotsDone
         ]