+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 1.0.0-beta-004
+
+- Bind.fragment deprecated, use Bind.el instead
+- Documentation code examples have max height, with faded text and Exand/Collapse buttons
 ### 1.0.0-beta-003
 
 - Implemented "Open in REPL" in docs (App)